movements and docstrings
Removed old file actions.py
Moved DBAuthContext to context.py
Moved base modules (logger, context, routes) to their own folder
Added documentation for context and working with routes
