+++ src/base_modules/routes.py
@@ -1,5 +1,6 @@
 """
 В этом файле описана логика работы с путями, которые мы храним для каждого пользователя
+NO IMPORTS IN THIS FILE
 """
 DEFAULT_ROUTE = '/'  # Корневое значение, когда пользователь не зашёл ни в какуб команду
