+++ src/entrypoint.py
@@ -1,5 +1,5 @@
 import telebot
-from src.constants import global_context
+from src.base_modules.context import global_context
 
 
 def run_bot(message):
