+++ src/execute_decorator.py
@@ -2,7 +2,7 @@
 В этом файле описан декоратор для функций, работающих при вызове бота
 """
 from functools import wraps
-from src.logger import Logger
+from src.base_modules.logger import Logger
 
 
 def message_execute_decorator(logger: Logger, on_error):
